+++ NEWS.md
@@ -1,6 +1,10 @@
 
 # vctrs (development version)
 
+* `vec_rbind()` and `vec_c()` with data frame inputs now consistently
+  preserve the names of list-columns, df-columns, and matrix-columns
+  (#689).
+
 * The internal version of `vec_assign()` now has support for assigning
   names and inner names. For data frames, the names are assigned
   recursively.